ARM: mmp: code strengthenment based on coverity
kernel upgrade from .35 to 3.0;

+++ include/linux/mfd/88pm860x.h
@@ -568,13 +568,13 @@ struct pm860x_backlight_pdata {
 	int		id;
 	int		pwm;
 	int		iset;
-	unsigned long	flags;
+	int		flags;
 };
 
 struct pm860x_led_pdata {
 	int		id;
 	int		iset;
-	unsigned long	flags;
+	int		flags;
 };
 
 struct pm860x_rtc_pdata {
